refactor: split the settings store
The 645-line module becomes stores/settings/: a core (payload, busy/
confirm machinery, dialog state), the organization domain, the profile
domain and the hash-routing dialog driver, composed in index.ts behind
the exact useSettingsModal(context) facade every page script already
spreads — verified key-for-key against the old return.
